Update speakbar as requested
